Reorganize the python code
