+++ .circleci/config.yml
@@ -10,7 +10,7 @@ jobs:
           name: Build Python packages
           no_output_timeout: 1.0h
           command: |
-            ./lemon/scripts/circleci/dockcross-manylinux-build-wheels.sh cp27 cp35 cp36 cp37
+            ./lemon/scripts/circleci/dockcross-manylinux-build-wheels.sh cp35 cp36 cp37
       - store_artifacts:
           path: dist
           destination: dist
