+++ CMakeLists.txt
@@ -93,3 +93,4 @@ if (${LEMON_BUILD_PROGS})
 endif()
 
 add_subdirectory(lang)
+
