+++ lang/python/tests/tmscore.py
@@ -4,7 +4,7 @@ class MyWorkflow(lemon.Workflow):
     def __init__(self):
         import lemon
         lemon.Workflow.__init__(self)
-        self.native = lemon.open_file("../../test/files/1AAQ.mmtf")
+        self.native = lemon.open_file("../../../test/files/1AAQ.mmtf")
     def worker(self, entry, pdbid):
         import lemon
         junk = lemon.PositionVec()
